Fall back to the packet based frame at the uplink
It seams that the CC1120 can not handle well frames without preamble.
Due to the G3RUH scrambler, we do not have a repeated alternating
sequence to use it as a preamble.

For now, we will continue with the packet based scheme for the uplink.
The downlink supports the legacy 9600 FSK modem flawlessly though.

The uplink could possibly support it too, but the clock recovery should
be implemented on the MCU